Refactor test utils
