+++ lhotse/testing/dummies.py
